Initialize project with first stab at card_data generation
Got "CRLF will be replaced by LF" warning a lot--not sure if I need to address
it but I'm gonna go ahead. No actual index file, yet, but I've begun working
on the first ideas for how to use the psuedoclassical approach for generating
all the cards I'll need as objects. Lots of questions remain! Huzzah!